requirements.txt
openai>=1.0.0
textual>=0.47.0
rich>=13.0.0
cryptography>=41.0.0
PyYAML>=6.0
ruamel.yaml>=0.18.0
tenacity>=8.2.0
elevenlabs>=1.0.0
faster-whisper>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
responses>=0.25.0
httpx>=0.25.0
